Revert "Connection pooler for replica (#1127)"
This reverts commit 1314ff2a74fff376cfbb6f43fe13c7d47bf2a2c6.
