+++ charts/postgres-operator/crds/postgresqls.yaml
@@ -190,8 +190,6 @@ spec:
                 type: string
               enableConnectionPooler:
                 type: boolean
-              enableReplicaConnectionPooler:
-                type: boolean
               enableLogicalBackup:
                 type: boolean
               enableMasterLoadBalancer:
