+++ manifests/complete-postgres-manifest.yaml
@@ -18,8 +18,7 @@ spec:
     - createdb
   enableMasterLoadBalancer: false
   enableReplicaLoadBalancer: false
-  enableConnectionPooler: false # enable/disable connection pooler deployment
-  enableReplicaConnectionPooler: false # set to enable connectionPooler for replica service
+#  enableConnectionPooler: true  # not needed when connectionPooler section is present (see below)
   allowedSourceRanges:  # load balancers' source ranges for both master and replica services
   - 127.0.0.1/32
   databases:
@@ -127,19 +126,18 @@ spec:
 #  - 01:00-06:00  #UTC
 #  - Sat:00:00-04:00
 
-# overwrite custom properties for connection pooler deployments
-#  connectionPooler:
-#    numberOfInstances: 2
-#    mode: "transaction"
-#    schema: "pooler"
-#    user: "pooler"
-#    resources:
-#      requests:
-#        cpu: 300m
-#        memory: 100Mi
-#      limits:
-#        cpu: "1"
-#        memory: 100Mi
+  connectionPooler:
+    numberOfInstances: 2
+    mode: "transaction"
+    schema: "pooler"
+    user: "pooler"
+    resources:
+      requests:
+        cpu: 300m
+        memory: 100Mi
+      limits:
+        cpu: "1"
+        memory: 100Mi
 
   initContainers:
   - name: date
