+++ manifests/postgresql.crd.yaml
@@ -186,8 +186,6 @@ spec:
                 type: string
               enableConnectionPooler:
                 type: boolean
-              enableReplicaConnectionPooler:
-                  type: boolean
               enableLogicalBackup:
                 type: boolean
               enableMasterLoadBalancer:
