+++ pkg/cluster/database.go
@@ -473,8 +473,8 @@ func (c *Cluster) execCreateOrAlterExtension(extName, schemaName, statement, doi
 }
 
 // Creates a connection pool credentials lookup function in every database to
-// perform remote authentication.
-func (c *Cluster) installLookupFunction(poolerSchema, poolerUser string, role PostgresRole) error {
+// perform remote authentification.
+func (c *Cluster) installLookupFunction(poolerSchema, poolerUser string) error {
 	var stmtBytes bytes.Buffer
 
 	c.logger.Info("Installing lookup function")
@@ -567,11 +567,12 @@ func (c *Cluster) installLookupFunction(poolerSchema, poolerUser string, role Po
 			failedDatabases = append(failedDatabases, dbname)
 			continue
 		}
+
 		c.logger.Infof("pooler lookup function installed into %s", dbname)
 	}
 
 	if len(failedDatabases) == 0 {
-		c.ConnectionPooler[role].LookupFunction = true
+		c.ConnectionPooler.LookupFunction = true
 	}
 
 	return nil
